Try Command, Attributes, Skills, Leveling
Added keywords for some terms, so that if needed they can be easily changed.

Added keywords for card terms to keep consistent naming/spelling.

Moved difficulty scale into a story card that's loaded on init.

Added common args for meta info.

Added config story card that is loaded on init. Settings include defualt dc, check dice, show for rolls and exp, and exp for actions.

Added attributes and skill story cards that are loaded into lists on init.

Implemented the try command which takes an attribute of skill keyword from the argText and runs a check using the players stats. Then appends the result to input and any info to message.

Changed the return of commandExtract to include parsed args and meta. (Avoids having to do it as a seperate step in the input function)

Removed the isDice logic from the typeChecker, as it's not used or needed atm.

Changed parseArgs to take the args as the full commandEntry is not needed.

Changed how optional and require args are defined and added a new format for args to have names. Args can now have one or more types, with the first type being the preferred.

Parsed args now parses the JSON of matched cards, so that it doesn't need to be done by commands.

matchArgs() now handles multiple types for an argument in a command.

Tightened the naming of custom types with the TRPG prefix, and some titles too.

Updated the character template to have level, attributes and skills loaded into it. Also load, update, and newChar all handle attributes and skills now.

Added new function to help validate if a character exists by checking all their cards.

Various functions for leveling & experience calculations, and one for attribute mod.

Updated the universalParser.js to be a full mock of the entire project.