Inventory Fixes
Mocked addStoryCards and updateStoryCards for the sandbox; to help check story card persistence.

Fixed numbers parsing as strings in the argsParser process.

Reverted loadCharacter and added an if case to use the template or story card to load data, for certain sections.

The comeback of updateCharSection to reduce how often every section was unnecessarily being updated.

Adjusted messages for take and drop; and added a strict case to drop, to prevent valid characters dropping items they do not have, or more than they have.