Take & Drop Many Things
Overhauled the parser and command specs to allow for sets of types/args to be defined. This allows commands like take and drop to specify multiples of a single or set of args.

Updated doTake and doDrop to allow for multiple items in one command. Amended drop text slightly.

Fixed a bug in doTry that incorrectly handled attribute mods.