Trade Command
Added trade command to take and remove items in one command.

Fixed a bug with sets not consuming tokens, thus reading the previous sets on subsequent sets.

Isolated the inventory add/remove logic of take and drop into helper functions to be used in trade and any other inventory management.