Inventory Actions
Added a new command for inventory action which use the +/- symbols to identify obtained and removed items.

Added memoryArea to config, so that users can choose which context area the instructions will go. (future proofing)

Added a clear action log call to the init method. This will make it easier to check if the AI actually produced a new result block.

Added a config help text for the config card entry. Moved the config data into the description of the card.

Added a inventory field to the character card, using the same format as skills.

Renamed skillTemplate into listTempData and created a reverse function listTemplateString.

Unified tryInstructions into actionInstructions for general use across any AI driven command. Same for saveActionResult.